permission

+++ lib/ui/screen/history_view.dart
@@ -2,12 +2,14 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:insta_downloader/enums/post_availability_enum.dart';
+import 'package:insta_downloader/enums/status_enum.dart';
 import 'package:insta_downloader/models/file_info_model.dart';
 import 'package:insta_downloader/ui/widget/history_template.dart';
 import 'package:insta_downloader/utils/dialogue_helper.dart';
 import 'package:insta_downloader/utils/downloader.dart';
 import 'package:insta_downloader/utils/file_checker.dart';
 import 'package:insta_downloader/utils/method_channel.dart';
+import 'package:insta_downloader/utils/permission.dart';
 import 'package:insta_downloader/utils/reponse_helper.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -31,6 +33,7 @@ class _HistoryViewState extends State<HistoryView> {
     });
   }
 
+  //TODO Optimise
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -48,6 +51,10 @@ class _HistoryViewState extends State<HistoryView> {
   }
 
   void popUpMenuFunction(String value, int index) async {
+    if (await getSdk() < 29 && !(await getDownloadPermission())) {
+      responseHelper(context, Status.PERMISSION_NOT_GRANTED);
+      return;
+    }
     Map check = await checkAllFiles(list[index]);
     PostAvailability postAvailability = check['post_availability'];
     List<String> availableFiles = check['available_files_uri'];
@@ -67,6 +74,10 @@ class _HistoryViewState extends State<HistoryView> {
         Clipboard.setData(ClipboardData(text: list[index].description));
         break;
       case 'delete':
+        if (await getSdk() < 29 && !(await getDownloadPermission())) {
+          responseHelper(context, Status.PERMISSION_NOT_GRANTED);
+          return;
+        }
         deleteFiles(availableFiles);
         //fire and forget
         DatabaseHelper.instance.delete(list[index]);
